requirements_dev.txt
#
# This file is autogenerated by pip-compile
# Make changes in requirements_dev.in, then run this to update:
#
#    pip-compile requirements_dev.in
#
alabaster==0.7.7          # via sphinx
babel==2.2.0              # via sphinx
colorama==0.3.6           # via pytest, sphinx
django-appconf==1.0.1
docutils==0.12            # via sphinx
flake8==2.5.2
isort==4.2.2
jinja2==2.8               # via sphinx
markupsafe==0.23          # via jinja2
mccabe==0.4.0
pep257==0.7.0
pep8-naming==0.3.3
pep8==1.7.0               # via flake8
py==1.4.31                # via pytest
pyenchant==1.6.6
pyflakes==1.0.0           # via flake8
pygments==2.1             # via sphinx
pytest-django==2.9.1
pytest==2.8.7
pytz==2015.7              # via babel
selenium==2.51.1
six==1.10.0               # via django-appconf, sphinx, sphinxcontrib-spelling
snowballstemmer==1.2.1    # via sphinx
sphinx-rtd-theme==0.1.9   # via sphinx
sphinx==1.3.5
sphinxcontrib-spelling==2.1.2
